+++ src/wsjcpp_core.cpp
@@ -1,3 +1,29 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2019-2026 Evgenii Sopov
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+*/
+
+// original source-code: https://github.com/wsjcpp/wsjcpp-core
+
 #include "wsjcpp_core.h"
 
 #ifndef _MSC_VER
@@ -5,7 +31,7 @@
     #include <sys/time.h>
     #include <unistd.h>
     #include <arpa/inet.h>
-#else 
+#else
     #include <direct.h>
     #define PATH_MAX 256
 #endif
@@ -1297,3 +1323,36 @@ bool WsjcppResourcesManager::extractFiles(const std::string &sWorkspace) {
 }
 */
 
+namespace wsjcpp {
+
+const std::string &Core::englishAlphabetLowerCase() {
+    static const std::string ret = "abcdefghijklmnopqrstuvwxyz";
+    return ret;
+}
+
+const std::string &Core::englishAlphabetUpperCase() {
+    static const std::string ret = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
+    return ret;
+}
+
+const std::string &Core::englishAlphabetBothCase() {
+    static const std::string ret = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
+    return ret;
+}
+
+const std::string &Core::englishAlphabetBothCaseAndNumbers() {
+    static const std::string ret = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+    return ret;
+}
+
+std::string Core::randomString(const std::string &alphabet, int length) {
+    std::string ret = "";
+    ret.resize(length, alphabet[0]);
+    int alphabet_len = alphabet.length();
+    for (int i = 0; i < length; i++) {
+        ret[i] = alphabet[rand() % alphabet_len];
+    }
+    return ret;
+}
+
+} // namespace wsjcpp